"Added hidden input fields for Web3Forms submission: subject, redirect, and from\_name"
This commit adds three new hidden input fields to the contact form on the index.html page. These fields are used to specify the subject of the form submission, the URL to redirect to after submission, and the name of the sender, respectively. The values for these fields are set to "New Form Submission", "<https://marsstrong.vercel.app/success.html>", and "Marsstrong", respectively.